requirements.txt
numpy
NREL-rex
pytest
sphinx==4.4.0
myst-parser
pyyaml
pandas
flake8
matplotlib
